requirements.txt
# requirements.txt
Flask
psycopg2-binary
python-dotenv
Flask-Session
Flask-Cors